SSP-2009 Fix on/off-plan display corruption issues
Reported issue involved loading a MAP template, navigating to a
different tool, coming back to MAP, and the display of the
student's on/off plan status would be corrupt, e.g. just the
"question mark" icon.

The most proximate cause of the display being unexpected in general
was that neither dismissing the MAP tool nor loading it back up
reset the current MAP plan singleton. And because the component
that looks up and displays the student's on/off-plan status
initializes before the component that loads the "current" plan
upon navigating to the MAP tool, the former thought it was
dealing with a template, not a plan, so it gave up on calculating
the status. But when it did so it only partially hid the view
components that represent that status.

The actual patch ended up being rather large for several reasons,
esp because the current plan needed to be reset in several locations
and to avoid regressing to a state where we were loading the plan
status several times on every plan render. The patch also ended up
simplifying the plan lookup plumbing code in MapPlanToolViewController
to eliminate all the nested functions for tracking multiple service
calls. It just wasn't necessary in this case and was making debugging
more difficult than it needed to be. Also tried to be more consistent
w/r/t when that component marks its view as loading and finished
loading.