SSP-2015 Clear dirty flag when resetting current Plan state
Not sure why, but in some cases the dirty flag was being set on the
singleton "current plan" - seemingly just when navigating from a user
with a plan to a user without a plan after having visited the MAP tool
for the former. My theory is that the navigation to the plan-less user
would either skip the dirty check for some reason or set the dirty
flag itself. And because the target user didn't herself have a plan,
some or all of the plan response mapping code would be skipped
altogether and the flag wouldn't be unset. So the next time you
navigated away from the planless user it would look like the singleton
plan was dirty.

Worrisome that I can't find the root cause, but clearing the dirty
flag every time the singleton plan model is reset seems to fix the
symptoms.