SSP-2017 Exit person sync gracefully on empty person result set
This behavior should have been present regardless, but the root cause
of the empty result sets in the originally reported issue was the
presence of a soft-deleted person record. PersonDao.getAll() does not
take the requested object status filter into account when producing
its count for the unpaginated result set, but it does when actually
running invidual paginated queries. So when a soft-deleted person
exists, the total count did not match the count that could actually
be returned. Thus the sync job would request an additional page
which would always be empty.

The PersonDao.getAll() behavior has been around since April 2012, so
I did not make any changes to it for this ticket. That will need to
be dealt with separately.