SSP-2003 Use different stores for caseload and bulk reassign
Previously both the caseload "search" view and the bulk reassignment
view used the same store for holding the "current" caseload. This
led to unexpected effects in the former view when the latter would
load b/c the first thing it would do is wipe out the current contents
of the store.