SSP-2264 Output read-only coach attr in Caseload Assignment
Expanded DB query projection underlying the /person/coach API
to include all the fields supported by the client side model.
Changed the return type on that API controller method to preserve
those fields in the TO. And modified the client-side form to output
these values into read-only form fields, all of which had the
"sync'd" label removed (if present) since that no longer makes
sense as a warning to the end user that local changes might be
over-written.