SSP-2241 Initialize SecurityContext before running background jobs
As described in patched javadoc, this works around a memory leak
caused by excessive SspUser creation in our Hibernate flush
interceptor when trying to look up the "auditable" current user.
Each of the those SspUser initializations ended up adding a
reference to a ThreadLocal list, which, for a "auditable-intensive"
job like external person sync, resulted in explosive growth in
list size and then GC thrash leading up to an OOM.

Conflicts:
	src/main/java/org/jasig/ssp/service/impl/ScheduledTaskWrapperServiceImpl.java

This was originally a cherry-pick of e060adbd708f74d13d5b64c31384bf2688d64622.
The conflict was a trivial imports discrepancy.