SSP-2289 Support additional EA-targeted SSO params
SSP-side support for several new params attached to EA-targeted
"Signed URL SSO" requests"

  - sectionCode
  - studentUserName
  - termCode
  - schoolId

The schoolId support is actually on the SSP-Platform side, so
this commit doesn't directly address that param. And this commit
doesn't make a whole lot of sense without those supporting
SSP-Platform commits.

Special thanks to Stanly CC for sponsoring the patches on which
this commit was originally based.