SSP-2654 Honor DOB search filters prior to epoch
DOB search filters prior to epoch were being ignored b/c
the presence of the filter was checked for by comparing
its value to zero. Since a date is effectively a int with
zero being the epoch any values prior to or on the epoch
were considered equivalent to an empty field and ignored.