SSP-2657 Validate directory search form before submission
Intended specifically to pick up bad DOB values before they're
submitted. As such that field was altered a bit in a few
other ways... in particular it is now validated as you type
and the date format mask was tweaked slightly to use 'y'
instead of 'c' for the century fields and was moved inside
the form field itself to match the convention in the rest
of the form.

Unfortunately I could not find a way to make the validation
message match a human-readable date format. So the field's
default text is mm/dd/yyyy but if you try to submit something
not in that format, you'll get a message telling to to
submit something in m/d/Y.