SSP-2710 Expire persistent prog status before creating new
Transitioning a person's program status requires a SQL update
to expire the old and an insert to create the new. A trigger
fires after each to update mv_directory_person. If the
insert runs first, the trigger will error out b/c the person
has two active program statuses. Hibernate has no idea that
we have a business rule here, and the two program status
records are completely unrelated to each other from Hib's
perspective. So it makes no guarantees about the order of the
insert and update when the next session flush occurs. In
some or all cases we were seeing the insert run first.

This patch forces the session to flush after the update.