SSP-2245 Change labels for Task Link to disallow HTML
This commit handles the client-side ExtJS portion of SSP-2245 and changes the labels for Task Referral Link in all the places it appears to discourage HTML markup. In the admin view the label was additionally modified to mention to administrators to remove HTML markup if it's present in existing links.