SSP-2245 Fix Link Display To Be Schema Independent
This commit changes the Link display in the Action Plan Add and Edit Task UI to be schema independent and robust for all valid and a few invalid URLs.
It also changes the task subreport to follow the UI changes, although it departs from the consistency of the UI due to jasperreports constraints.
Finally, this fix adds another tooltip on the Link edit in the Tasks.grid itself when a row is double-clicked.