SSP-2732 Fix student full name sorting in caseload reassignment
Not 100% sure why this patch fixes it. All I did was remove the
view code that used Ssp.util.ColumnRendererUtils.renderStudentFullName
as the 'Name' column renderer in both caseload reassignment forms.
We didn't really need that code for actual on-screen formatting, so
it was nice to be able to remove it. And b/c it was introduced
as part of previous SSP-2732 commits, I deleted that function
completely.