SSP-2832 Widen message.sent_* fields
Very similar to a patch with the same description in 2.6.0+,
but with support for the message_archive table removed.

Actually, most of that cleanup work was done in the 2.5.3
patch on which this patch was based. This patch just needed
additional, 2.4.x-specific masterChangeLog.xml conflict
resolution b/c 2.4.x has/needs so many fewer changelogs.

Conflicts:
	src/main/java/org/jasig/ssp/model/ArchivedMessage.java
	src/main/resources/org/jasig/ssp/database/masterChangeLog.xml

Conflicts:
	src/main/resources/org/jasig/ssp/database/masterChangeLog.xml