SSP-2920 Fix program status buttons for Faculty and Support Staff
This isn't limited to Faculty or Support Staff and really had to do with the
 view behaviour on which of the 3 tabs (My Caseload, WatchList, and Search)
 a user had. For example this fixes a case where WatchList and Search were the
 only two tabs a user had access to, that just didn't work before this patch.
It also fixes very much on the surface the case where a user doesn't have access
 to search, but has access to the other tab(s).
In general, this fixes assumptions of only two tabs or assumptions that were ignoring the Watchlist case.
 This does still leave the behaviour of assuming Search is not hidden, but as I mentioned above, a surface level
 fix to SearchTab removes the security issue and forces Search to not appear thus "fixing" it for now.
Fixing all of that above had the side-effect of fixing the Program Status buttons. This is because
 the correct tabs are now set as Active, before the code to execute returns nothing if the tab wasn't active.