SSP-2994 Fix MAP Issues in UI
Fixes 2 browser errors one for ShowMapStatus undefined for the MapStatus
pop-up and the other an Ext.JSON decode error for the MapStatusReports store
when that comes back as empty.

Also fixes the layout issues of the 'Filter By' field being cut-off and a
collapsed object causing a speck to appear on the screen near the 'Filter By'.
An adjustment was also made to the large Map Plan Status Calculation pop-up
to better fit height convention as well as the Email Plan pop-up.