NOJIRA Fix MAP Display Issues with Missing External Term
Relates to main cause in SSP-2980.
This fixes a situation that really shouldn't occur, but due to SSP-3016,
an issue was found with MAP not loading. This occurred if a term wasn't
present in the external_term table, but was present in a MAP Plan.
Guard code was added so that at least the MAP would display and function
normally for terms that are available in the external_term table.