SSP-3018 Avoid permanently disabling Ext layouts when Main interrupted
An attempt at a perf optimization when rendering evaluated success
indicators in Main would potentially disable Ext layouts for the
entire app if that rendering failed in any way.

Layouts were being suspended at the beginning of
EvaluatedSuccessIndicatorGroupsController.renderViewSpecs() and
re-enabled at the end. But if the view had already been
destroyed, the rendering operations before the re-enablement
would fail and the re-enablement would not be invoked. This
left the UI completely crippled.

The solution is just to take the optimization out. It probably
wasn't helping much in the first place.