SSP-3129 Fix EA Response Status and Details to Display
Re-factors or changes in EarlyAlert controllers, models and views client-side must have caused this issue. The EarlyAlert service was using an OutcomeStore that wasn't loaded previously by anything. Furthermore, the calculation to display Status was using a nodetype and fields non-existant in the underlying models. This commit fixes those two issues to once-again display EA Response Status and Details in the grid in the Early Alert tool.