SSP-2579 Add Option to Route All Email to Single Address
Added to ssp-config properties a new field:

reroute_all_mail_to_address=

When an email address is entered all mail will be re-routed to that address.
If left blank as is in the default ssp-config file, then this configuration
is disabled. When enabled, the original recipient's, cc's and bcc's addresses
will be added along with a small message to the top of the email body
 for test/qa purposes.