SSP-3157 Fix Add/Edit a Student refreshing DirectoryPerson
The person table trigger on INSERT/UPDATE didn't work when editing a
 student or performing an UPDATE query.
 The materiazlied view (mv_directory_person) would never update and
the newly changed user couldn't be found in Search among other possibilities.

Upon investigation it was found that the person table on UPDATE/INSERT has
 3 scenarios to consider:
    1. Insert new person/user/student into person table where school_id does not exist in external_person
    2. Insert new person/user/student into person where school_id exists in external_person
    3. Update an existing person record

This commit updates the triggers to reflect those 3 scenarios and was tested
to work properly in those scenarios. More general testing is needed since
 person table is core to SSP. SQL Server should also get additional scrutiny
 since the trigger was changed to a MERGE to perform an 'UPSERT' function.