SSP-3158 Modify MAP Plan Status Calculation to Consider Last Retaken Course
Previous commits on this ticket, the MAP Plan Status Calculation was modified to look for a course with a passing grade in the transcript when the course had a failing grade.  Additionally threw configuration, the school can choose to only consider the course was passed if the last time the course has a passing grade.  The configuration option is map_plan_status_use_any_passing_course.  If true, any passing grade is considered passing.  If false, only a passing grade for the last time the course was taken is considered passing.