SSP-3155 Bind course-specific store when rendering elective chooser
Fixes issue where the elective chooser would always be bound
to the electives for the "last" elective-choosable course in the
relevant semester. That was happening b/c the store binding
occurred in the cell renderer, but the editor (combobox) itself
was scoped to the column. This moves the store binding down to
the cell level.