SSP-3197 Add risk counts to Watchload/Caseload/Search
Adds a count of low/medium configured risks to the Caseload/Watchlist/Search results.

Searching by students that have a configured medium/low risk is nearly done from a back-end perspective. The field will need to be added to the SearchForm front-end and request object.

Generating early alerts on configured success indicators moving to low is also nearly complete. The functionality to create the alert in early alert service needs to be created or re-use existing methods. This is trickier since we don't keep success indicator state. One thought is to keep a count of low alerts and then when a new configured indicator goes to low; create a new EA or add a Response if one for success indicators already exists.