SSP-3230 Fix SSG to lower size of JDBC Parameters
This only fixes External SSG to minimize the occurence of hitting an error in JTDS with more than 2000 parameter markers. So, far that has been the main culprit in causing this issue, although anytime having more than 2000 of a reference item could cause this issue. A user selecting more than 2000 parameters also is almost impossible to fix. Queries can be broken up, but due to sorting and paging everything needs to be combined, sorted and paged.